Refactors steps to an easier to work with organization
- groups checks and actions
- one statement per file
- adds error handling for ElementNotFound
