+++ behave.ini
@@ -1,6 +1,8 @@
 [behave]
 format=formatters.cucumber_json:PrettyCucumberJSONFormatter
-outfiles=reports/cucumber.json
+outfiles=/app/veripy/reports/cucumber.json
 show_snippets=false
 quiet=true
 show_timings=false
+paths=/app/veripy
+verbose=true
