+++ veripy/steps/aggregates/__init__.py
@@ -1,0 +1,1 @@
+from aggregates.actions import * # noqa
