+++ veripy/steps/capture/capture_content.py
@@ -1,0 +1,22 @@
+import logging
+from behave import then
+
+logger = logging.getLogger('capture')
+
+
+@then('write the "{element}" generated by the system')
+def capture_element_text(context, element):
+    """ Given an element, record the value of the element for later display
+    in the report.
+
+    ::
+
+        Then write the "{element}" generated by the system.
+
+    """
+    logger.info(f'Capturing content from  "{element}"')
+    try:
+        text = context.page[element].text
+    except context.page.ElementNotFound:
+        raise AssertionError(f'The {element} was not found on the page')
+    context.step.stored_value = text
