+++ veripy/steps/capture/capture_screenshot.py
@@ -1,15 +1,12 @@
 import logging
-
 from behave import then
-
 from veripy.utils.browsers import screenshot_bytes
 
-
-logger = logging.getLogger('screenshots')
+logger = logging.getLogger('capture')
 
 
 @then('take a screen shot')
-def take_screen_shot(context):
+def capture_screenshot(context):
     """ Have the browser take a screenshot of the current window. """
     logger.info('Taking a screenshot.')
     context.step.screenshots.append(screenshot_bytes(context))
