+++ veripy/steps/content/__init__.py
@@ -1,0 +1,2 @@
+from content.actions import * # noqa
+from content.checks import * # noqa
