+++ veripy/steps/navigation/__init__.py
@@ -1,0 +1,2 @@
+from navigation.actions import * # noqa
+from navigation.checks import * # noqa
